FIX KABOOM SYNTAX ERROR
Revert back to an older version of Kaboom to prevent syntax errors in
the bundled code. This error exists only in DuckDuckGo (not Chrome or Safari).
This error does not occur for Kaboom versions prior to 3000.1.0. A github
issue was opened for this error (https://github.com/replit/kaboom/issues/812).