SEPARATE CREATE AND SHOW UI
We now first create invisible UI so that their position can be used
to place/attach future elements. This is only possible when the UI is
created before other components. Then we can use 'showUI' in development
mode to showcase where the UI components are placed on screen. Move file
from utils to entities because it is now always used.