CONVERT TO VITE AND TYPESCRIPT
We continued to have issues with bundling when trying to import
new plugins (kaboom-flex-ui). It was recommended by kaboom developers
to switch to using `vite` as a bundler instead of `webpack`. We also
converted to typescript to prevent persisting issues and prevent issues
with proving incorrect data types to kaboom functions.